resolve conflict in test.js
