+++ lib/tests/run_tests.js
@@ -2,8 +2,7 @@ const async = require('async');
 const fs = require('fs-extra');
 const Mocha = require('mocha');
 const path = require('path');
-const Test = require('./test.js');
-const utils = require('../utils/utils.js');
+const Test = require('./test');
 
 function getFilesFromDir(filePath, cb) {
   fs.readdir(filePath, (err, files) => {
@@ -49,24 +48,20 @@ module.exports = {
         });
       },
       function setupGlobalNamespace(next) {
-        // ---------------- Deprecated code ------------------------------------------------------------
         global.assert = require('assert');
 
-        global.config = function(config) {
-          configOptions = utils.recursiveMerge(configOptions, config);
-        };
-        // TODO: check how to pass the options
-        //global.EmbarkSpec = new Test(options);
+        // TODO put default config
+        const test = new Test();
+        global.embark = test;
+        global.config = test.config.bind(test);
 
         // TODO: this global here might not be necessary at all
-        global.EmbarkSpec = new Test({});
-        global.web3 = global.EmbarkSpec.web3;
+        global.web3 = global.embark.web3;
 
         global.contract = function(describeName, callback) {
           return Mocha.describe(describeName, callback);
         };
         next();
-        // ---------------- Deprecated code ------------------------------------------------------------
       }
     ], (err) => {
       if (err) {
