+++ package.json
@@ -40,12 +40,12 @@
     "ejs": "^2.5.8",
     "eth-ens-namehash": "^2.0.8",
     "eth-lib": "^0.2.8",
-    "ethereumjs-testrpc": "^6.0.3",
     "ethereumjs-wallet": "^0.6.0",
     "file-loader": "^1.1.5",
     "finalhandler": "^1.1.1",
     "follow-redirects": "^1.2.4",
     "fs-extra": "^2.0.0",
+    "ganache-cli": "^6.1.0",
     "globule": "^1.1.0",
     "hard-source-webpack-plugin": "^0.6.6",
     "http-shutdown": "^1.2.0",
